+++ openNASR/__init__.py
@@ -6,6 +6,8 @@ from .airspace import ArtccBoundary as ArtccBoundary
 from .airspace import ArtccRepository as ArtccRepository
 from .airspace import ClassAirspace as ClassAirspace
 from .airspace import ClassAirspaceRepository as ClassAirspaceRepository
+from .airspace import Maa as Maa
+from .airspace import MaaRepository as MaaRepository
 from .atc import AtcFacility as AtcFacility
 from .atc import AtcFacilityRepository as AtcFacilityRepository
 from .atc import Radar as Radar
@@ -38,6 +40,10 @@ from .plotting import plot_airspace as plot_airspace
 from .plotting import plot_airport_procedures as plot_airport_procedures
 from .records import ArtccRecord as ArtccRecord
 from .records import ClassAirspaceRecord as ClassAirspaceRecord
+from .records import MaaRecord as MaaRecord
+from .records import MaaContactRecord as MaaContactRecord
+from .records import MaaRemarkRecord as MaaRemarkRecord
+from .records import MaaShapePointRecord as MaaShapePointRecord
 from .records import AirwayRecord as AirwayRecord
 from .records import AirwaySegmentRecord as AirwaySegmentRecord
 from .records import AtcFacilityRecord as AtcFacilityRecord
@@ -106,6 +112,12 @@ __all__ = [
     "HoldingPatternRemarkRecord",
     "HoldingPatternRepository",
     "HoldingPatternSpeedAltitudeRecord",
+    "Maa",
+    "MaaContactRecord",
+    "MaaRecord",
+    "MaaRemarkRecord",
+    "MaaRepository",
+    "MaaShapePointRecord",
     "MilitaryOperation",
     "MilitaryOperationRecord",
     "MilitaryOperationRepository",